+++ hypertunity/tests/test_utils.py
@@ -1,5 +1,4 @@
 import queue
-from typing import Union
 
 import pytest